[refactor] Move arch files
