+++ taichi/backends/arch.h
@@ -4,7 +4,6 @@
 #include "taichi/common/core.h"
 
 namespace taichi {
-namespace lang {
 
 enum class Arch : int {
 #define PER_ARCH(x) x,
@@ -29,5 +28,4 @@ bool arch_use_host_memory(Arch arch);
 
 int default_simd_width(Arch arch);
 
-}  // namespace lang
 }  // namespace taichi
